Adding configuration
