+++ app/navigation/LinkingConfiguration.ts
@@ -1,13 +1,7 @@
-/**
- * Learn more about deep linking with React Navigation
- * https://reactnavigation.org/docs/deep-linking
- * https://reactnavigation.org/docs/configuring-links
- */
-
-import { LinkingOptions } from '@react-navigation/native';
 import * as Linking from 'expo-linking';
+import { LinkingOptions } from '@react-navigation/native';
 
-import { RootStackParamList } from '../interfaces';
+import { RootStackParamList } from '../types/navigation';
 
 const linking: LinkingOptions<RootStackParamList> = {
   prefixes: [Linking.createURL('/')],
@@ -15,9 +9,9 @@ const linking: LinkingOptions<RootStackParamList> = {
     screens: {
       Root: {
         screens: {
-          TabOne: {
+          Home: {
             screens: {
-              TabOneScreen: 'one'
+              Home: 'one'
             }
           },
           TabTwo: {
