+++ app/screens/Home.tsx
@@ -1,0 +1,22 @@
+import { Text, View } from '../components/Themed';
+import { StyleSheet } from 'react-native';
+
+export default function Home() {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Just for today</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center'
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: 'bold'
+  }
+});
