+++ app/screens/ModalScreen.tsx
@@ -1,8 +1,7 @@
-import { StatusBar } from 'expo-status-bar';
 import { Platform, StyleSheet } from 'react-native';
-
-import EditScreenInfo from '../components/EditScreenInfo';
 import { Text, View } from '../components/Themed';
+import EditScreenInfo from '../components/EditScreenInfo';
+import { StatusBar } from 'expo-status-bar';
 
 export default function ModalScreen() {
   return (
