+++ app/types/component.d.ts
@@ -1,0 +1,9 @@
+interface IExternalLink {
+  children: JSX.Element;
+  href: string;
+  style: {};
+}
+
+interface IEditScreenInfo {
+  path: string;
+}
