+++ app/types/content.d.ts
@@ -10,13 +10,3 @@ interface IDailyReflection {
   source: string;
   reflection: string;
 }
-
-interface IExternalLink {
-  children: JSX.Element;
-  href: string;
-  style: {};
-}
-
-interface IEditScreenInfo {
-  path: string;
-}
