+++ babel.config.js
@@ -1,7 +1,6 @@
 module.exports = (api) => {
   api.cache(true);
   return {
-    presets: ['babel-preset-expo'],
-    plugins: [require.resolve('expo-router/babel')]
+    presets: ['babel-preset-expo']
   };
 };
