+++ package.json
@@ -15,38 +15,39 @@
     "test": "jest"
   },
   "dependencies": {
+    "expo": "~48.0.9",
     "expo-asset": "~8.9.1",
     "expo-constants": "~14.2.1",
     "expo-font": "~11.1.1",
     "expo-linking": "~4.0.1",
+    "expo-router": "^1.3.0",
     "expo-splash-screen": "~0.18.1",
     "expo-status-bar": "~1.4.4",
     "expo-web-browser": "~12.1.1",
-    "expo": "~48.0.9",
+    "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-native": "0.71.4",
     "react-native-safe-area-context": "4.5.0",
     "react-native-screens": "~3.20.0",
-    "react-native-web": "0.18.12",
-    "react-native": "0.71.4",
-    "react": "18.2.0"
+    "react-native-web": "0.18.12"
   },
   "devDependencies": {
     "@babel/core": "^7.21.3",
     "@expo/vector-icons": "^13.0.0",
     "@react-navigation/bottom-tabs": "^6.5.7",
-    "@react-navigation/native-stack": "^6.9.12",
     "@react-navigation/native": "^6.1.6",
+    "@react-navigation/native-stack": "^6.9.12",
     "@types/jest": "^29.5.0",
+    "@types/react": "~18.0.29",
     "@types/react-native": "~0.71.5",
     "@types/react-test-renderer": "^18.0.0",
-    "@types/react": "~18.0.29",
     "@typescript-eslint/eslint-plugin": "^5.56.0",
     "@wildpastry/eslint-config": "^1.6.4",
+    "eslint": "^8.36.0",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
-    "eslint": "^8.36.0",
-    "jest-expo": "~48.0.2",
     "jest": "^29.5.0",
+    "jest-expo": "~48.0.2",
     "prettier": "^2.8.7",
     "react-test-renderer": "18.2.0",
     "ts-node": "^10.9.1",
